Create LICENSE
All scripts are written to a CD/DVD/USB or any portable media and self postage mailed via a local government sanctioned mail service for the soul purpose of contesting dates related to authoring of the file within a reasonably short period of time after publishing publicly published scripts. If you can present a reasonable argument to me that you're the (within reason) original author to a script segment, I am happy to republish the script with appropriate modifications to help protect your rights as an intellectual property owner.